requirements.txt
streamlit
torch
torchvision
Pillow
requests
numpy
matplotlib
seaborn